requirements.txt
# requirements.txt
opencv-python
numpy
Pillow
pyyaml
pandas
easyocr
streamlit
matplotlib
shapely
rich

